Backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiofiles==23.2.1
    # via gradio
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via
    #   datasets
    #   fsspec
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   gradio
    #   httpx
    #   starlette
attrs==24.2.0
    # via
    #   aiohttp
    #   lancedb
cachetools==5.5.0
    # via lancedb
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dataclasses-json==0.6.7
    # via langchain-community
datasets==2.20.0
    # via -r requirements.in
decorator==4.4.2
    # via
    #   moviepy
    #   retry
deprecation==2.1.0
    # via lancedb
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
fastapi==0.114.0
    # via gradio
ffmpy==0.4.0
    # via gradio
filelock==3.16.0
    # via
    #   datasets
    #   huggingface-hub
fonttools==4.53.1
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.5.0
    # via
    #   datasets
    #   gradio-client
    #   huggingface-hub
gradio==4.39.0
    # via -r requirements.in
gradio-client==1.1.1
    # via gradio
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via
    #   gradio
    #   gradio-client
    #   langsmith
huggingface-hub==0.24.6
    # via
    #   datasets
    #   gradio
    #   gradio-client
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
imageio==2.35.1
    # via moviepy
imageio-ffmpeg==0.5.1
    # via moviepy
importlib-resources==6.4.4
    # via gradio
jinja2==3.1.4
    # via gradio
joblib==1.4.2
    # via
    #   pynndescent
    #   scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
kiwisolver==1.4.7
    # via matplotlib
lancedb==0.8.2
    # via -r requirements.in
langchain==0.2.4
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.2.4
    # via -r requirements.in
langchain-core==0.2.6
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.2.1
    # via
    #   -r requirements.in
    #   langchain
langsmith==0.1.116
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
llvmlite==0.43.0
    # via
    #   numba
    #   pynndescent
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   gradio
    #   jinja2
marshmallow==3.22.0
    # via dataclasses-json
matplotlib==3.9.1
    # via
    #   -r requirements.in
    #   gradio
    #   seaborn
mdurl==0.1.2
    # via markdown-it-py
moviepy==1.0.3
    # via -r requirements.in
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
mypy-extensions==1.0.0
    # via typing-inspect
numba==0.60.0
    # via
    #   pynndescent
    #   umap-learn
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   datasets
    #   gradio
    #   imageio
    #   langchain
    #   langchain-community
    #   matplotlib
    #   moviepy
    #   numba
    #   opencv-python
    #   pandas
    #   pyarrow
    #   pylance
    #   scikit-learn
    #   scipy
    #   seaborn
    #   umap-learn
opencv-python==4.10.0.82
    # via -r requirements.in
orjson==3.10.7
    # via
    #   gradio
    #   langsmith
overrides==7.7.0
    # via lancedb
packaging==24.1
    # via
    #   datasets
    #   deprecation
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   lancedb
    #   langchain-core
    #   marshmallow
    #   matplotlib
pandas==2.2.2
    # via
    #   -r requirements.in
    #   datasets
    #   gradio
    #   seaborn
pillow==10.3.0
    # via
    #   -r requirements.in
    #   gradio
    #   imageio
    #   matplotlib
predictionguard==2.1.0
    # via -r requirements.in
proglog==0.1.10
    # via moviepy
py==1.11.0
    # via retry
pyarrow==15.0.0
    # via
    #   datasets
    #   pylance
pyarrow-hotfix==0.6
    # via datasets
pydantic==2.9.0
    # via
    #   fastapi
    #   gradio
    #   lancedb
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.23.2
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.18.0
    # via rich
pylance==0.12.1
    # via lancedb
pynndescent==0.5.13
    # via umap-learn
pyparsing==3.1.4
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
python-multipart==0.0.9
    # via gradio
pytube==15.0.0
    # via -r requirements.in
pytubefix==6.5.2
    # via -r requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via
    #   datasets
    #   gradio
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
ratelimiter==1.2.0.post0
    # via lancedb
requests==2.32.3
    # via
    #   datasets
    #   huggingface-hub
    #   lancedb
    #   langchain
    #   langchain-community
    #   langsmith
    #   moviepy
    #   predictionguard
    #   youtube-transcript-api
retry==0.9.2
    # via lancedb
rich==13.8.0
    # via typer
ruff==0.6.4
    # via gradio
scikit-learn==1.5.1
    # via
    #   -r requirements.in
    #   pynndescent
    #   umap-learn
scipy==1.14.1
    # via
    #   pynndescent
    #   scikit-learn
    #   umap-learn
seaborn==0.13.2
    # via -r requirements.in
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.34
    # via
    #   langchain
    #   langchain-community
starlette==0.38.5
    # via fastapi
tabulate==0.9.0
    # via predictionguard
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
threadpoolctl==3.5.0
    # via scikit-learn
tomlkit==0.12.0
    # via gradio
tqdm==4.66.5
    # via
    #   datasets
    #   huggingface-hub
    #   lancedb
    #   moviepy
    #   proglog
    #   umap-learn
typer==0.12.5
    # via gradio
typing-extensions==4.12.2
    # via
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via
    #   pandas
    #   pydantic
umap-learn==0.5.6
    # via -r requirements.in
urllib3==2.2.2
    # via
    #   gradio
    #   requests
uvicorn==0.30.6
    # via gradio
websockets==11.0.3
    # via gradio-client
webvtt-py
    # via -r requirements.in
xxhash==3.5.0
    # via datasets
yarl==1.11.0
    # via aiohttp
youtube-transcript-api==0.6.2
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
